requirements.txt
asgiref==3.4.1
Django==3.2.8
django-filter==21.1
djangorestframework==3.12.4
Markdown==3.3.4
Pillow==8.3.2
pytz==2021.3
sqlparse==0.4.2
validate-docbr==1.8.2
